Update crop interface to allow cropping of remote files
